+++ server/models/accountModels.go
@@ -6,8 +6,8 @@ type Account struct {
 	ID        int       `json:"id"`
 	FirstName string    `json:"firstName"`
 	LastName  string    `json:"lastName"`
-	Number    int64     `json:"number"`
+	Number    int       `json:"number"`
 	Password  string    `json:"password"`
-	Balance   int64     `json:"balance"`
+	Balance   int       `json:"balance"`
 	CreatedAt time.Time `json:"createdAt"`
 }
